[spaces.fv.default] moved out of playground, refs #15
